+++ arithmoi.cabal
@@ -48,6 +48,8 @@ library
                           , semigroups >= 0.8 && < 1.0
 
     exposed-modules     : Math.NumberTheory.ArithmeticFunctions
+                          Math.NumberTheory.ArithmeticFunctions.Class
+                          Math.NumberTheory.ArithmeticFunctions.Standard
                           Math.NumberTheory.Logarithms
                           Math.NumberTheory.Moduli
                           Math.NumberTheory.MoebiusInversion
@@ -70,6 +72,7 @@ library
                           Math.NumberTheory.Primes.Testing
                           Math.NumberTheory.Primes.Testing.Certificates
                           Math.NumberTheory.Primes.Heap
+                          Math.NumberTheory.UniqueFactorization
     other-modules       : Math.NumberTheory.Utils
                           Math.NumberTheory.Unsafe
                           Math.NumberTheory.Logarithms.Internal
